+++ tests/testthat/test-interval2state.R
@@ -1,0 +1,80 @@
+test_that("interval2state works as expected", {
+  #create a interval dataset
+  states <- tibble::tibble(Datetime = c("2023-08-15 6:00:00",
+                                        "2023-08-15 23:00:00",
+                                        "2023-08-16 6:00:00",
+                                        "2023-08-16 22:00:00",
+                                        "2023-08-17 6:30:00",
+                                        "2023-08-18 1:00:00",
+                                        "2023-08-18 6:00:00",
+                                        "2023-08-18 22:00:00",
+                                        "2023-08-19 6:00:00",
+                                        "2023-08-19 23:00:00",
+                                        "2023-08-20 6:00:00",
+                                        "2023-08-20 22:00:00"),
+                           State = rep(c("wake", "sleep"), 6),
+                           Source = 
+                             rep(c("Participant", "Environment"), each = 6)) %>% 
+    dplyr::group_by(Source)
+  intervals <- sc2interval(states)
+  
+  #create a dataset with states
+  states <-
+    sample.data.environment %>% dplyr::group_by(Source) %>%
+    interval2state(State.interval.dataset = intervals,
+                   ID.colname.dataset = Source,
+                   ID.colname.interval = Source) %>% 
+    dplyr::slice(6000, 28000) %>% dplyr::pull(State)
+  
+  expect_equal(states,c(NA, "wake", "sleep"))
+  
+})
+
+test_that("interval2state throws errors as expected", {
+  #create a interval dataset
+  states <- tibble::tibble(Datetime = c("2023-08-15 6:00:00",
+                                        "2023-08-15 23:00:00",
+                                        "2023-08-16 6:00:00",
+                                        "2023-08-16 22:00:00",
+                                        "2023-08-17 6:30:00",
+                                        "2023-08-18 1:00:00",
+                                        "2023-08-18 6:00:00",
+                                        "2023-08-18 22:00:00",
+                                        "2023-08-19 6:00:00",
+                                        "2023-08-19 23:00:00",
+                                        "2023-08-20 6:00:00",
+                                        "2023-08-20 22:00:00"),
+                           State = rep(c("wake", "sleep"), 6),
+                           Id = "Participant")
+  intervals <- sc2interval(states)
+  
+  #create a dataset with states
+  expect_error(interval2state("Data", State.interval.dataset = intervals,
+                              ID.colname.dataset = Source),
+               "dataset is not a dataframe")
+  
+  expect_error(interval2state(dataset = sample.data.environment,
+                              State.interval.dataset = intervals,
+                              ID.colname.dataset = Source2),
+               "ID.colname.dataset must be part of the dataset")
+  
+  expect_error(interval2state(dataset = sample.data.environment,
+                              Interval.colname = Interval2,
+                              State.interval.dataset = intervals,
+                              ID.colname.dataset = Source),
+               "Interval.colname must be part of the State.interval.dataset")
+  
+  expect_error(interval2state(dataset = sample.data.environment,
+                              State.interval.dataset = intervals,
+                              ID.colname.interval = Source,
+                              ID.colname.dataset = Source),
+               "ID.colname.interval must be part of the State.interval.dataset")
+  
+  expect_error(interval2state(dataset = sample.data.environment,
+                              State.interval.dataset = sample.data.environment,
+                              State.colname = Source,
+                              ID.colname.dataset = Source,
+                              ID.colname.interval = Source,
+                              Interval.colname = Datetime))
+  
+})